Fixed compilation warnings

+++ Zend/zend_arena.h
@@ -59,11 +59,11 @@ static zend_always_inline void* zend_arena_alloc(zend_arena **arena_ptr, size_t 
 
 	size = ZEND_MM_ALIGNED_SIZE(size);
 
-	if (EXPECTED(size <= (arena->end - ptr))) {
+	if (EXPECTED(size <= (size_t)(arena->end - ptr))) {
 		arena->ptr = ptr + size;
 	} else {
 		size_t arena_size = 
-			UNEXPECTED((size + ZEND_MM_ALIGNED_SIZE(sizeof(zend_arena))) > (arena->end - (char*) arena)) ?
+			UNEXPECTED((size + ZEND_MM_ALIGNED_SIZE(sizeof(zend_arena))) > (size_t)(arena->end - (char*) arena)) ?
 				(size + ZEND_MM_ALIGNED_SIZE(sizeof(zend_arena))) :
 				(arena->end - (char*) arena);
 		zend_arena *new_arena = (zend_arena*)emalloc(arena_size);
